draft_gen/requirements.txt
anthropic
openai
fastapi
uvicorn[standard]
pydantic
python-dotenv
